+++ app/build.gradle
@@ -18,8 +18,8 @@ android {
         applicationId "com.seok.gfd"
         minSdkVersion 26
         targetSdkVersion 28
-        versionCode 6
-        versionName "2.16.6"
+        versionCode 7
+        versionName "2.17.7"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
