+++ app/src/main/java/com/seok/gfd/views/LauncherActivity.kt
@@ -21,23 +21,6 @@ class LauncherActivity : AppCompatActivity() {
         window.setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN)
 
         initSetUI()
-
-//        val scalablelayout = ScalableLayout(this,commonUtils.getScreenWidth(), commonUtils.getScreenHeight())
-//        scalablelayout.backgroundColor = Color.LTGRAY
-//        val tv = TextView(this)
-//        tv.text = "testjfalskdflasjdlkfjaskldfjlksadjfklsajdflksjdlkfjaslkdfjlasdflasdjflkasdf"
-//        tv.backgroundColor = Color.YELLOW
-//        scalablelayout.addView(tv, 10f, 10f, commonUtils.getScreenWidth()/2, commonUtils.getScreenHeight())
-//        scalablelayout.setScale_TextSize(tv, commonUtils.getScreenHeight()/10)
-
-//        scalablelayout.setScale_TextSize(tv, 20f)
-//        val imageVIew = ImageView(this)
-//        imageVIew.setImageResource(R.drawable.gfd_logo)
-//        scalablelayout.addView(imageVIew, 200f, 30f, 50f, 50f)
-
-//        launcher_const_layout.addView(scalablelayout)
-
-
         startLoading()
     }
 
@@ -47,7 +30,7 @@ class LauncherActivity : AppCompatActivity() {
             startActivity(intent)
             overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
             finish()
-        }, 100)
+        }, 1500)
     }
 
     private fun initSetUI(){
