--- app/src/main/res/layout/fragment_main2.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".views.MainFragment">
-
-
-    <com.ssomai.android.scalablelayout.ScalableLayout
-            android:id="@+id/main_top_scalable_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:scale_base_height="250"
-            app:scale_base_width="1080"
-            tools:ignore="MissingPrefix">
-
-        <ImageView
-                android:id="@+id/main_image_profile"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/gfd_logo"
-                app:scale_height="200"
-                app:scale_left="30"
-                app:scale_top="30"
-                app:scale_width="200" />
-
-
-        <TextView
-                android:id="@+id/main_tv_user_name"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/spoqa_han_sans_bold"
-                android:textColor="@color/colorText"
-                android:text="HelloGfd"
-                app:scale_height="100"
-                app:scale_left="270"
-                app:scale_textsize="55"
-                app:scale_top="40"
-                app:scale_width="500" />
-
-        <TextView
-                android:id="@+id/main_tv_today"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/spoqa_han_sans_thin"
-                android:gravity="right"
-                android:textColor="@color/colorText"
-                android:text="2020-01-01"
-                app:scale_height="100"
-                app:scale_left="750"
-                app:scale_textsize="35"
-                app:scale_top="20"
-                app:scale_width="300" />
-
-        <ImageView
-                android:id="@+id/main_image_location"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/icon_bio"
-                app:scale_height="50"
-                app:scale_left="270"
-                app:scale_top="145"
-                app:scale_width="50" />
-
-        <TextView
-                android:id="@+id/main_tv_user_bio"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/spoqa_han_sans_light"
-                android:text="Selfie"
-                android:textColor="@color/colorText"
-                app:scale_height="70"
-                app:scale_left="330"
-                app:scale_textsize="35"
-                app:scale_top="135"
-                app:scale_width="610" />
-
-    </com.ssomai.android.scalablelayout.ScalableLayout>
-    <com.ssomai.android.scalablelayout.ScalableLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:scale_base_height="1920"
-            app:scale_base_width="1080"
-            tools:ignore="MissingPrefix">
-        <com.ogaclejapan.smarttablayout.SmartTabLayout
-                android:id="@+id/main_tab_smart_layout"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                app:stl_indicatorAlwaysInCenter="false"
-                app:stl_indicatorWithoutPadding="false"
-                app:stl_indicatorInFront="false"
-                app:stl_indicatorInterpolation="smart"
-                app:stl_indicatorGravity="bottom"
-                app:stl_indicatorColor="@color/stCommit"
-                app:stl_indicatorThickness="3dp"
-                app:stl_indicatorWidth="auto"
-                app:stl_indicatorCornerRadius="2dp"
-                app:stl_overlineColor="@color/fiCommit"
-                app:stl_overlineThickness="0dp"
-                app:stl_underlineColor="@color/colorTopBar"
-                app:stl_underlineThickness="1dp"
-                app:stl_dividerColor="@color/colorTopBar"
-                app:stl_dividerThickness="1dp"
-                app:stl_defaultTabBackground="?attr/selectableItemBackground"
-                app:stl_defaultTabTextAllCaps="true"
-                app:stl_defaultTabTextColor="@color/colorText"
-                app:stl_defaultTabTextSize="15sp"
-                app:stl_defaultTabTextHorizontalPadding="10dp"
-                app:stl_defaultTabTextMinWidth="120sp"
-                app:stl_distributeEvenly="false"
-                app:stl_clickable="true"
-                app:stl_titleOffset="24dp"
-                app:stl_drawDecorationAfterTab="false"
-                app:scale_height="150"
-                app:scale_top="250"
-                app:scale_width="1080"/>
-
-        <androidx.viewpager.widget.ViewPager
-                android:id="@+id/main_view_pager"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_below="@id/main_tab_smart_layout"
-                app:scale_height="1520"
-                app:scale_top="400"
-                app:scale_width="1080"/>
-    </com.ssomai.android.scalablelayout.ScalableLayout>
-</FrameLayout>
